Refactoring + Working of handle_quote and handle_dollar
The aforomentioned functions need to rewrite the data->input, so we'll
no longer have another pointer for it (it would be a mess for freeing
the memory if the input had multiple quotes and dollar signs)
As such, some refactoring of parse_input was in need.
Now, working on the aforomentioned functions.