Escaping chars is implemented!!
Still need to deal with single escape at the end of the line (new line
char) and with semicommas. The escaping also works for pipes and
ampersands, but only if adding a couple of lines to the executioner that
were already added on the main_builtin branch (which will be merged
here).

What is missing:
  -Deal with nl and semicommas.
  -Implement on the execution part the && and ||
  -Redirections
  -Append
  -Heredoc
  -Signals
  -Norm friendly code